requirements.txt
fastapi
uvicorn
pandas
python-multipart
pyarrow
httpx
python-docx
pydantic
numpy
prophet
python-multipart
python-dotenv
reportlab
pywin32
scikit-learn
